Rename --override-channels to --ignore-channels (#178)
The command line flag `--override-channels` is overloaded.

The conda cli uses the `--override-channels` flag along with the
`--channels` flag in order to allow conda to configure channels
per run. In conda, users must specify the `--channels` cli arg
if the `--override-channels` arg is provided.

In conda-pypi the existing `--override-channels` flag means to
ignore the configured channels and just use the local pypi package
cache.

These two ideas are not the same, so they should not have the same
name.